+++ programs/drift/src/math/fulfillment/tests.rs
@@ -134,7 +134,7 @@ mod determine_perp_fulfillment_methods {
         assert_eq!(
             fulfillment_methods,
             [
-                PerpFulfillmentMethod::Match(Pubkey::default(), 0),
+                PerpFulfillmentMethod::Match(Pubkey::default(), 0, 99 * PRICE_PRECISION_U64),
                 PerpFulfillmentMethod::AMM(None)
             ]
         );
@@ -208,7 +208,7 @@ mod determine_perp_fulfillment_methods {
             fulfillment_methods,
             [
                 PerpFulfillmentMethod::AMM(Some(maker_order.price)),
-                PerpFulfillmentMethod::Match(Pubkey::default(), 0),
+                PerpFulfillmentMethod::Match(Pubkey::default(), 0, 101 * PRICE_PRECISION_U64),
                 PerpFulfillmentMethod::AMM(None)
             ]
         );
@@ -278,9 +278,9 @@ mod determine_perp_fulfillment_methods {
         assert_eq!(
             fulfillment_methods,
             [
-                PerpFulfillmentMethod::Match(Pubkey::default(), 0),
+                PerpFulfillmentMethod::Match(Pubkey::default(), 0, 99 * PRICE_PRECISION_U64),
                 PerpFulfillmentMethod::AMM(Some(101 * PRICE_PRECISION_U64)),
-                PerpFulfillmentMethod::Match(Pubkey::default(), 1),
+                PerpFulfillmentMethod::Match(Pubkey::default(), 1, 101 * PRICE_PRECISION_U64),
                 PerpFulfillmentMethod::AMM(None),
             ]
         );
@@ -354,8 +354,12 @@ mod determine_perp_fulfillment_methods {
         assert_eq!(
             fulfillment_methods,
             [
-                PerpFulfillmentMethod::Match(Pubkey::default(), 0),
-                PerpFulfillmentMethod::Match(Pubkey::default(), 1),
+                PerpFulfillmentMethod::Match(Pubkey::default(), 0, 99 * PRICE_PRECISION_U64),
+                PerpFulfillmentMethod::Match(
+                    Pubkey::default(),
+                    1,
+                    99 * PRICE_PRECISION_U64 + PRICE_PRECISION_U64 / 2
+                ),
                 PerpFulfillmentMethod::AMM(None),
             ]
         );
@@ -426,9 +430,9 @@ mod determine_perp_fulfillment_methods {
             fulfillment_methods,
             [
                 PerpFulfillmentMethod::AMM(Some(102 * PRICE_PRECISION_U64)),
-                PerpFulfillmentMethod::Match(Pubkey::default(), 0),
+                PerpFulfillmentMethod::Match(Pubkey::default(), 0, 102 * PRICE_PRECISION_U64),
                 PerpFulfillmentMethod::AMM(Some(103 * PRICE_PRECISION_U64)),
-                PerpFulfillmentMethod::Match(Pubkey::default(), 1),
+                PerpFulfillmentMethod::Match(Pubkey::default(), 1, 103 * PRICE_PRECISION_U64),
                 PerpFulfillmentMethod::AMM(None),
             ]
         );
@@ -498,9 +502,9 @@ mod determine_perp_fulfillment_methods {
         assert_eq!(
             fulfillment_methods,
             [
-                PerpFulfillmentMethod::Match(Pubkey::default(), 0),
+                PerpFulfillmentMethod::Match(Pubkey::default(), 0, 101 * PRICE_PRECISION_U64),
                 PerpFulfillmentMethod::AMM(Some(99 * PRICE_PRECISION_U64)),
-                PerpFulfillmentMethod::Match(Pubkey::default(), 1),
+                PerpFulfillmentMethod::Match(Pubkey::default(), 1, 99 * PRICE_PRECISION_U64),
                 PerpFulfillmentMethod::AMM(None),
             ]
         );
@@ -570,8 +574,8 @@ mod determine_perp_fulfillment_methods {
         assert_eq!(
             fulfillment_methods,
             [
-                PerpFulfillmentMethod::Match(Pubkey::default(), 0),
-                PerpFulfillmentMethod::Match(Pubkey::default(), 1),
+                PerpFulfillmentMethod::Match(Pubkey::default(), 0, 102 * PRICE_PRECISION_U64),
+                PerpFulfillmentMethod::Match(Pubkey::default(), 1, 101 * PRICE_PRECISION_U64),
                 PerpFulfillmentMethod::AMM(None),
             ]
         );
@@ -642,9 +646,9 @@ mod determine_perp_fulfillment_methods {
             fulfillment_methods,
             [
                 PerpFulfillmentMethod::AMM(Some(99 * PRICE_PRECISION_U64)),
-                PerpFulfillmentMethod::Match(Pubkey::default(), 0),
+                PerpFulfillmentMethod::Match(Pubkey::default(), 0, 99 * PRICE_PRECISION_U64),
                 PerpFulfillmentMethod::AMM(Some(98 * PRICE_PRECISION_U64)),
-                PerpFulfillmentMethod::Match(Pubkey::default(), 1),
+                PerpFulfillmentMethod::Match(Pubkey::default(), 1, 98 * PRICE_PRECISION_U64),
                 PerpFulfillmentMethod::AMM(None),
             ]
         );