+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
+- program: apply 10bps to protected market maker limit orders ([#1417](https://github.com/drift-labs/protocol-v2/pull/1417))
 - program: allow lighthouse at end of swap ([#1429](https://github.com/drift-labs/protocol-v2/pull/1429))
 
 ### Fixes